+++ views/container/iframe.php
@@ -1,11 +1,17 @@
 <?php
 
 use humhub\helpers\Html;
+use humhub\modules\custom_pages\models\CustomPage;
+
+/* @var CustomPage $page */
+/* @var string $navigationClass */
+/* @var string $url */
 
 $cssClass = ($page->hasAttribute('cssClass') && !empty($page->cssClass)) ? $page->cssClass : 'custom-pages-page';
 ?>
 
 <iframe class="<?= Html::encode($cssClass) ?>" id="iframepage" <?= $page->iframe_attrs ?? '' ?>
+        aria-label="<?= Html::encode($page->title) ?>"
         style="width:100%; height: 100%; min-height: 400px;"
         src="<?= Html::encode($url); ?>"></iframe>