+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 1.12.10 (Unreleased)
 -------------------------
 - Enh #489: New reserved element "parent"
+- Enh #493: Add attribute "aria-label" to iframes
 
 1.12.9 (December 1, 2025)
 -------------------------