+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ Changelog
 -------------------------
 - Enh #489: New reserved element "parent"
 - Enh #492: New filter visibility filter for Mobile App users
+- Enh #493: Add attribute "aria-label" to iframes
+- Fix #494: Fix double section "Advanced" of the template element "File Download"
 
 1.12.9 (December 1, 2025)
 -------------------------